refs #9 ui modularization
