+++ caclulator/starter/lib/main.dart
@@ -1,6 +1,6 @@
 
 import 'package:calculator_modularization_presentation_starter/di/provider/calculator.dart';
-import 'package:calculator_modularization_presentation_starter/ui/screen/calculator.dart';
+import 'package:calculator_modularization_ui/ui.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
