+++ caclulator/starter/ui/LICENSE
@@ -1,0 +1,1 @@
+TODO: Add your license here.
