+++ caclulator/starter/ui/lib/ui.dart
@@ -1,0 +1,3 @@
+library ui;
+
+export 'src/ui.dart';
